tmf: Fix value of time offset in XY chart viewer being wrong for a while
Right after calling setWindowStartTime/setWindowEndTime, the value of
getTimeOffset was wrong until setTimeOffset was called later, in a data
building thread for example.

Since the time offset is deduced from the start time (startTime - 1), we can
remove this field and setter. This means that the time offset will always be
consistent with the start time. Similarly, the window duration can be deduced
(startTime - endTime) therefore we can remove this setting and field as well.

An example of how this bug can manifest itself is inside a resize listener.
If both windowStartTime and timeOffset are used to calculate the position of a
value on the chart, since the timeOffset can be inconsistent, the position
would be wrong.

Change-Id: If59a9d2c0dfc24245949f8ba693dd89e6498820f
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>