Tmf.Ui: fix Multithreaded correctness - Call to static DateFormat
The problem has been very well explained here:
http://fahdshariff.blogspot.ca/2010/08/dateformat-with-multiple-threads.html
As this code may be used in multi-threaded environments, not making it thread
safe will result in sporadically mangled dates. This patch fixes it by
synchronizing each access with the list of all items. More performance would
be had by using a threadlocal scheme where setting the timezone resets the
structure, or having specific locks for each formatter, but the legibility of
the class would greatly suffer.

Change-Id: If8e9b118b3db15a2a0c088d4ebb2fa7c20839885
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>