Tmf.Ui: fix Multithreaded correctness - Call to static DateFormat
The problem has been very well explained here:
http://fahdshariff.blogspot.ca/2010/08/dateformat-with-multiple-threads.html
As this code may be used in multi-threaded environments, not making it thread
safe will result in sporadically mangled dates. This patch fixes it by storing
the date formats in ThreadLocals, each thread will have a local copy of the final
object.

Change-Id: If8e9b118b3db15a2a0c088d4ebb2fa7c20839885
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>