tmf: Allow negative time range selection
This patch allows the selection end time to be lower than the selection
start time, better representing the user selection.

It will help to correctly update the end boundary using atomic
operations (keyboard or button) that extend a selection to the left of
the selection start.

The status line message delta is updated to allow a negative value.

Change-Id: I5228441cb049487783d4b0fdcba7ef766077677f
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>