common.test: change test to use an Integer queue instead of a char queue
Tests with a char queue are limited to 26 possible values (a-z)
This patch generates a list of integers which are increasing numbers, so
each element is unique and easy to infer what the next one is. This allows
more rigourous tests with larger datasets. For the moment the test runs
40000 elements instead of 200. Also, this allows easier debugging as values
do not repeat.

Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Change-Id: I45b495b82f6212c8aeb00b6d540253467e325965