ctf: Allow mixing unspecified stream IDs with ID 0
If an event does not specify a stream ID, we will assign it to
stream with ID 0. This means they will be put in the same stream
as the events specifically defining stream 0.

A corrolary is that if a stream does not specify it's id, it too
is assigned to 0.

However, the trace must not define any stream IDs other than 0.

This behaviour matches the one of Babeltrace.

Change-Id: I0628acaa8566dde0a3659fe56d5a8e903c6feb26
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>