ctf: Provide a Maven property to specify a ctf-testsuite commit to use
Instead of automatically pulling all changes from ctf-testsuite,
which may break our builds unexpectedly, we'll define a commit
to use in the plugin's pom.xml.

We can regularly update this pointer to pull in newer versions
of the test suite.

The property can also be overridden on the command-line, for
example using:

  mvn clean verify -DctfTestSuiteCommit=master

will use ctf-testsuite's master branch.

Change-Id: I06b8e11f12fec16d9c011f0f2a80b083fcadd8de
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>