tmf: Change updating of XY views to avoid incorrect time alignment
Before this patch, when updating the current window range the stored
window start and end time were updated before the corresponding x
values array were set in the SWT chart in a different thread. The
time aligment algorithm were querying the available alignment width
which used the new window start/end time to get the pixelCoordinate
for the start time. Since the corresponding x values hadn't been set
yet, an incorrect pixel coordinate was provided and incorrect
alignment width was calculated. This was just temporary, because
after updating the XY chart a new alignment was requested and then
the alignment was "fixed".

However, the temporary incorrect alignment is noticable by the user.

The patch only sets the window start and end time at the same time
when the x values are set.

Change-Id: I2a5be2430b19ee295360548e3fef62b2e04ba2a2
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>