tmf: Update calculation of offset and available width in XY chart
During window range updates the window start and end time of the
xy chart viewer used to be inconsistent with the corresponding
x-Series values. This lead to incorrect values when calculating
the point offset and available width for the time alignment during
the period of inconsistency. With this patch the offset and width
is calculated using the first and last time in the x-Series array.

Change-Id: I70eee0c08777c4b6797e2d6bbc41dd19fff05f64
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>