lttng: Disable spuriously failing Partial State System tests
The partial state system is in very bad shape and not used at all.
Until some energy is put into improving its implementation and
testing, it should not be depended upon.

To make matters worse, one of its tests is failing randomly on
Hudson. Since it is not in a supported state at the moment, I
suggest disabling these tests until we improve the PSS.

Change-Id: I592598670b2f2a122c8f638b264502a48dae177c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>