tmf: Bug 467751: Partial workaround for XY chart concurrency issue
This change will remove the usage of the TmfUIRefreshHandler which
delays the clearing of the chart.

However, it is still not thread-safe and it might fail. To fix it
correctly, the whole API of XY charts would have to looked at and
this won't be done for Mars.

Change-Id: Ib838864fd8fb7086391b44f0b3dbf0558617f5f2
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/48320
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Hudson CI
(cherry picked from commit 9b59296287a33916ae0855af3262491b3a4b2a13)