Add a new target to specify the API baseline
Eclipse 4.5 now offers specifying API baselines using target files.
This is much more simple than having to install and maintain a
separate Eclipse installation!

Add a new .target, under a baselines/ directory (to not confuse them
with "real" targets listing build dependencies) which we can use
for setting the API baseline.

It currently points to the latest Trace Compass build, which is
1.0 RC2, but could be updated once the 1.0 final is out. The APIs
should not change in between, in any case.

Change-Id: Ie83d6ee09de00d102a735da99cd50118c5efd1f5
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>