lttng: Split UST event names in a separate interface
Remove the definitions pertaining to event and field names in
UstMemoryStrings, and move them to a global interface.

These names can change from one version to the other of the
tracer, so this is the first step in abstracting away the actual
trace event names from the "concepts" the library and analysis
agree upon.

Change-Id: Ia8ccf528029cc8cd3e5de13e4e7041ad6e89eeae
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>