lttng: Introduce a debug-info analysis for UST traces
This UST-specific analysis will track the loading and unloading
of shared libraries, and will allow to match an instruction pointer
from an event context to the corresponding binary or library.

Once we have the correct binary (executable or .so) identified,
we can match it to an actual binary file with debug information,
and retrive the corresponding address/filename/line number, just
like is currently done with the Callstack view.

For now, the analysis will only map an instruction pointer to
the corresponding binary file. A future step will use CDT
binary parsing facilities to get the function name and line
number corresponding to the instruction pointer.

Change-Id: I6127e89cc5d3d8c7abff01610b8a6dc9a612f149
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>