swtbot: Fix ControlFlowViewTest
If the test host burps for a minimum of 400 ms between the 5th and the
6th click of the "Follow CPU Fwd" button, the test will fail because the
wait condition for the time range returns immediately since the 6th
selected event is a change of process that has the same timestamp as the
5th selected event and the listener notifier has timed out due to the
host delay, updating the trace manager with the selection time of the
5th selected event, and then the following assert on the tree selection
fails because the selected entry change event is still stuck in the new
listener notifier for another 400 ms.

A new condition is created to wait until the tree selection equals the
expected value, instead of using an assertion.

Additionally, the tests are updated to use the new click(int) with
modifier methods in SWTBot, allowing the test of range selection using
tool bar buttons while the SHIFT key is pressed.

Change-Id: Ic607d063b13248bd5ab91a87861ce3dad8a2b286
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>