tmf: Add color preferences for event raw viewer highlight and selection
The color settings can be changed in Window > Preferences > General >
Appearance > Colors and Fonts > Tracing.

Descriptions are added for all the Tracing font and color preferences.

The default trace event table font is changed to the system font instead
of the dialog font, which more accurately reflects the widget default.

Change-Id: Ia422655c2acd4b15b7ee9b01e07e80bbce5ea900
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>