ctf: Fix CtfIterator current event and location after seek
If an iterator has equal current and previous location (like it does
after the current event has been read once), the seek will not update
the current event even if it does not match the seek location. Then the
current event and current location of the iterator can return arbitrary
values that do not correspond with the seek location.

This can happen if the iterator instance is recycled for a different
context, which only happens when the iterator manager is full due to 100
concurrent contexts for the same trace, which probably can only happen
if there is a context leak, as the iterator is removed from the iterator
manager when it is disposed.

Change-Id: Ifd67b2c866955f424e11172b8e6fc27a35ad676a
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>