ctf: Fix incorrect location in CtfTmfContext after seek
After seeking by timestamp, the context location should be set to the
exact timestamp of the event, not the specified timestamp.

After seeking to the start of the trace, the context location should be
set to the exact timestamp of the first event, not to 0. This was a
regression introduced when the CtfLocation was made final.

Change-Id: I5d34653aede7afd5b5fef1baead532a8b1640fa7
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>