ctf: Fix location ratio handling in CtfTmfTrace
In getLocationRatio(), the provided location timestamp can be used
directly instead of requiring a context seek. Incidentally, the context
used in the current implementation was leaking as it was not disposed.

The total time range used for the ratio calculation should be based on
the trace start and end time (timestamp of first and last event). Using
the iterator range can result in Infinity ratio when the iterator has
only read a single event so far.

In seekEvent(double), the total time range used should also be based on
the trace start and end time. The trace getCurrentStartTime() and
getCurrentEndTime() methods use the packet start and end times which do
not necessarily match actual trace event timestamps.

Change-Id: Ic8bfcdebbd5c388a4048815d01124434e4c82ba6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>