ss: Introduce SegmentStore plugin
Unlike the State System, which assigns single states to attributes,
the Segment Store is a database interface for storing segments that
may overlap each other. It is closer in intent to a spatial database.

Just like state system backends, we provide an interface that can be
implemented in different ways. A first basic in-memory implementation
is provided.

Change-Id: I20a4a27b66ae005e8a22070752d8cc0a132a36f5
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>