lttng: Add a ILttngUstEventLayout
The event names exposed by LTTng-UST for the memory analysis have
changed in LTTng 2.7. They are now prefixed by "lttng_ust_*",
instead of the previous "ust_*".

To support various versions in parallel, we can use a "event
layout" interface, similarly to what was done for kernel traces
with IKernelEventLayout.

This patch adds a new ILttngUstEventLayout interface,
implementations for 2.6- and 2.7+ versions, and moves the
Memory Analysis to use this layout.

It also required removing the requirements from the memory
analysis, since the framework cannot currently check for event
names after a trace is initialized.

Fixes Bug 470186.

Change-Id: Id1f4bde4f5fc56bb3f6d003e51716944ce2c3f93
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>