tmf: Don't track event fields in ITmfEventType
ITmfEventType was tracking the fields of the events. This is
completely useless, because the fields are already tracked by the
ITmfEvent object already (with its getContent()), and nothing in
the framework actually uses the information stored by ITmfEventType.

We can remove these interface methods for a nice code cleanup.

ITmfEventType is now reduced to only tracking the type name.
Subclasses can still add more information, like BtfEventType does.

Change-Id: I33ba015a6eefa11cab6d3577e99e25e70b1b486d
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>