tmf: Use Strings in ITmfTraceWithPreDefinedEvents
Since ITmfEventType is now effectively just a String representing
the event name, we can use Strings directly and eventually remove
ITmfEventType altogether.

The first potential location is ITmfTraceWithPreDefinedEvents, which
returns the event types found in a trace. All places calling it were
wrapping it in the helper class's method to extract the event names
anyway, proving the uselessness of ITmfEventType once more.

This also revealed that CtfTmfTrace was taking great care of creating
arrays and lists to represents the event fields of each type, and
that these lists were never used.

Change-Id: I77cef7f41c4f1de781f1d30bad22369b7d3cf519
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>