tmf: Add support for blending sub-pixel events in time graph
This allows for many short time events mapping to the same pixel to be
blended in together instead of drawing only the first one and discarding
the remaining. This functionality is optional and off by default.

The Time Chart view is updated to use this scheme to allow the
highlighted ticks from a color settings filter to be noticeable even if
the time event is not the first event of that pixel.

Change-Id: I316f044157cef50f6744d128c3661964b97f09c1
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/51068
Reviewed-by: Hudson CI
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit 37a507dca88f92e040df4d860191a88224fca382)