tmf: Use getter methods instead of fields in TmfEvent#equals()
Since the various getter methods in TmfEvent are not final,
which means that subclasses can (and do) extend them, the
method calls should be used instead of the object's fields
to correctly compare two such objects.

TODO:
This change seems to expose some problems, where events that
were not technically equal were considered to be.

Change-Id: I2f4a2b8f455cfecd2b2a11f480092cd751eb50d3
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>